Postcode lottery
Added a postcode reggae, which may be suitable but will still let
through some non postcodes and needs more testing.

Added a new valit-len for character length add like so [no of char] |
[message].

Same for phone number, removed the need for hard coding in the jQuery
plugin. Add the length in the valit-tel att